Further progress, but still not finding a transaction with outputs and addresses

+++ priv/repo/migrations/20181221085001_create_inputs.exs
@@ -4,7 +4,7 @@ defmodule BlockChainExplorer.Repo.Migrations.CreateInputs do
   def change do
     create table(:inputs) do
       add :transaction_id, :integer
-      add :sequence, :integer
+      add :sequence, :bigint
       add :scriptsig, :string
       add :coinbase, :string
       add :asm, :string
